feat: initial release of module
This module is initially intended to be used for the development phase.
I feel much more comfortable knowing that the ec2 instances I am working
with have a unique keypair and I am unable to accidentally execute
commands on EC2 instances outside of the Terraform Code I am working on.